refactor: useInput 훅 제거에 따른 유효기간 업데이트
